\name{EWOC.Comb-package}
\alias{EWOC.Comb-package}
\alias{EWOC.Comb}
\docType{package}
\title{
\packageTitle{EWOC.Comb}
}
\description{
\packageDescription{EWOC.Comb}
}
\author{
\packageAuthor{EWOC.Comb}

Maintainer: \packageMaintainer{EWOC.Comb}
}
\references{Tighiouart M, Li Q and Rogatko A. A Bayesian adaptive design for estimating the maximuym tolerated dose curve using drug combinations in cancer phase I clinical trials. Statistics in Medicine. 2017, 36: 280-290.}
\keyword{ package }
